react router added

+++ frontend/src/components/Person.jsx
@@ -1,7 +1,19 @@
+import { useParams, useNavigate } from 'react-router-dom'
+
 const Person = ({ person, handleDeletePerson }) => {
+  const id = useParams().id
+  const navigate = useNavigate()
+
   const onDeleteClick = (e) => {
     e.preventDefault()
-    handleDeletePerson(person)
+    if (window.confirm(`Delete person "${person.name}"?`)) {
+      handleDeletePerson(person)
+      navigate('/persons')
+    }
+  }
+
+  if (!person) {
+    return null
   }
 
   return (
